Crate tendril

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Reexports

pub use fmt::Format;

pub use stream::TendrilSink;

Modules

fmt 	 [
Experimental
]
Marker types for formats.

stream 	 [
Experimental
]
Streams of tendrils.

Macros

format_tendril 	 [
Experimental
]
Create a StrTendril through string formatting.

Structs

Atomic 	 [
Experimental
]
A marker of an atomic (and hence concurrent) tendril.

IncompleteUtf8 	 [
Experimental
]

NonAtomic 	 [
Experimental
]
A marker of a non-atomic tendril.

SendTendril 	 [
Experimental
]
A simple wrapper to make Tendril Send.

Tendril 	 [
Experimental
]
Compact string type for zero-copy parsing.

Enums

SubtendrilError 	 [
Experimental
]
Errors that can occur when slicing a Tendril.

Traits

Atomicity 	 [
Experimental
]
The multithreadedness of a tendril.

ReadExt 	 [
Experimental
]
Extension trait for io::Read.

SliceExt 	 [
Experimental
]
Tendril-related methods for Rust slices.

Type Definitions

ByteTendril 	 [
Experimental
]
Tendril for storing binary data.

StrTendril 	 [
Experimental
]
Tendril for storing native Rust strings.