tendril::buf32

Struct Buf32

pub struct Buf32<H> {
    pub ptr: *mut H,
    pub len: u32,
    pub cap: u32,
}

A buffer points to a header of type H, which is followed by MIN_CAP or more bytes of storage.

Fields
ptr: *mut Hlen: u32cap: u32
Implementations

impl<H> Buf32<H>

pub unsafe fn with_capacity(cap: u32, h: H) -> Buf32<H>

pub unsafe fn destroy(self)

pub unsafe fn data_ptr(&self) -> *mut u8

pub unsafe fn data(&self) -> &[u8]

pub unsafe fn data_mut(&mut self) -> &mut [u8]

pub unsafe fn grow(&mut self, new_cap: u32)

Grow the capacity to at least new_cap.

This will panic if the capacity calculation overflows u32.