Struct tendril::fmt::UTF8

pub struct UTF8;

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Marker type for UTF-8 text.