Trait tendril::fmt::CharFormat

pub unsafe trait CharFormat<'a>: Format {
    type Iter: Iterator<Item = (usize, char)>;
    unsafe fn char_indices(buf: &'a [u8]) -> Self::Iter;

    fn encode_char<F>(ch: char, cont: F) -> Result<(), ()>
    where
        F: FnOnce(&[u8]);
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates a format which contains characters from Unicode (all of it, or some proper subset).

Associated Types

type Iter: Iterator<Item = (usize, char)>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterator for characters and their byte indices.

Required Methods

unsafe fn char_indices(buf: &'a [u8]) -> Self::Iter

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Iterate over the characters of the string and their byte indices.

You may assume the buffer is already validated for Format.

fn encode_char<F>(ch: char, cont: F) -> Result<(), ()> where
F: FnOnce(&[u8]),

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Encode the character as bytes and pass them to a continuation.

Returns Err(()) iff the character cannot be represented.