Trait tendril::fmt::SubsetOf

pub unsafe trait SubsetOf<Super>: Format where
    Super: Format, {
    fn revalidate_subset(x: &[u8]) -> bool { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates that one format is a subset of another.

The subset format can be converted to the superset format for free.

Provided Methods

fn revalidate_subset(x: &[u8]) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Validate the other direction of conversion; check if this buffer from the superset format conforms to the subset format.

The default calls Self::validate, but some conversions may implement a check which is cheaper than validating from scratch.