Struct tendril::stream::LossyDecoder

pub struct LossyDecoder<Sink, A = NonAtomic> where Sink: TendrilSink<UTF8, A>, A: Atomicity {
    // some fields omitted
}

A TendrilSink adaptor that takes bytes, decodes them as the given character encoding, lossily replace ill-formed byte sequences with U+FFFD replacement characters, and emits Unicode (StrTendril).

This allocates new tendrils for encodings other than UTF-8.

Methods

impl<Sink, A> LossyDecoder<Sink, A> where Sink: TendrilSink<UTF8, A>, A: Atomicity

fn new(encoding: EncodingRef, sink: Sink) -> LossyDecoder<Sink, A>

Create a new incremental decoder.

fn utf8(sink: Sink) -> LossyDecoder<Sink, A>

Create a new incremental decoder for the UTF-8 encoding.

This is useful for content that is known at run-time to be UTF-8 (whereas Utf8LossyDecoder requires knowning at compile-time.)

fn inner_sink(&self) -> &Sink

Give a reference to the inner sink.

fn inner_sink_mut(&mut self) -> &mut Sink

Give a mutable reference to the inner sink.