Struct tendril::stream::Utf8LossyDecoder

pub struct Utf8LossyDecoder<Sink, A = NonAtomic> where
    Sink: TendrilSink<UTF8, A>,
    A: Atomicity,  {
    pub inner_sink: Sink,
    // some fields omitted
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A TendrilSink adaptor that takes bytes, decodes them as UTF-8, lossily replace ill-formed byte sequences with U+FFFD replacement characters, and emits Unicode (StrTendril).

This does not allocate memory: the output is either subtendrils on the input, on inline tendrils for a single code point.

Fields
inner_sink: Sink

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<Sink, A> Utf8LossyDecoder<Sink, A> where
Sink: TendrilSink<UTF8, A>,
A: Atomicity,

fn new(inner_sink: Sink) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Create a new incremental UTF-8 decoder.