Trait tendril::stream::TendrilSink

pub trait TendrilSink<F, A = NonAtomic> where
    F: Format,
    A: Atomicity, {
    type Output;
    fn process(&mut self, t: Tendril<F, A>);

    fn error(&mut self, desc: Cow<'static, str>);

    fn finish(self) -> Self::Output;

    fn one<T>(self, t: T) -> Self::Output
    where
        Self: Sized,
        T: Into<Tendril<F, A>>,
    { ... }

    fn from_iter<I>(self, i: I) -> Self::Output
    where
        Self: Sized,
        I: IntoIterator,
        I::Item: Into<Tendril<F, A>>,
    { ... }

    fn read_from<R>(self, r: &mut R) -> Result<Self::Output>
    where
        Self: Sized,
        R: Read,
        F: SliceFormat<Slice = [u8]>,
    { ... }

    fn from_file<P>(self, path: P) -> Result<Self::Output>
    where
        Self: Sized,
        P: AsRef<Path>,
        F: SliceFormat<Slice = [u8]>,
    { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Trait for types that can process a tendril.

This is a "push" interface, unlike the "pull" interface of Iterator<Item=Tendril<F>>. The push interface matches html5ever and other incremental parsers with a similar architecture.

Associated Types

type Output

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

What the overall result of processing is.

Required Methods

fn process(&mut self, t: Tendril<F, A>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Process this tendril.

fn error(&mut self, desc: Cow<'static, str>)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates that an error has occurred.

fn finish(self) -> Self::Output

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates the end of the stream.

Provided Methods

fn one<T>(self, t: T) -> Self::Output where
Self: Sized,
T: Into<Tendril<F, A>>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Process one tendril and finish.

fn from_iter<I>(self, i: I) -> Self::Output where
Self: Sized,
I: IntoIterator,
I::Item: Into<Tendril<F, A>>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Consume an iterator of tendrils, processing each item, then finish.

fn read_from<R>(self, r: &mut R) -> Result<Self::Output> where
Self: Sized,
R: Read,
F: SliceFormat<Slice = [u8]>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Read from the given stream of bytes until exhaustion and process incrementally, then finish. Return Err at the first I/O error.

fn from_file<P>(self, path: P) -> Result<Self::Output> where
Self: Sized,
P: AsRef<Path>,
F: SliceFormat<Slice = [u8]>,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Read from the file at the given path and process incrementally, then finish. Return Err at the first I/O error.