Struct tendril::NonAtomic

pub struct NonAtomic(_);

A marker of a non-atomic tendril.

This is the default for the second type parameter of a Tendril and so doesn't typically need to be written.

This is akin to using Rc for reference counting.