Struct tendril::SendTendril

pub struct SendTendril<F> where
    F: Format,  { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A simple wrapper to make Tendril Send.

Although there is a certain subset of the operations on a Tendril that a SendTendril could reasonably implement, in order to clearly separate concerns this type is deliberately minimalist, acting as a safe encapsulation around the invariants which permit Sendness and behaving as an opaque object.

A SendTendril may be produced by Tendril.into_send() or SendTendril::from(tendril), and may be returned to a Tendril by Tendril::from(self).