Trait tendril::Atomicity

pub unsafe trait Atomicity: 'static { }

The multithreadedness of a tendril.

Exactly two types implement this trait:

Atomic: use this in your tendril and you will have a Send + Sync tendril which works across threads; this is akin to Arc.

NonAtomic: use this in your tendril and you will have a tendril which is neither Send nor Sync but should be a tad faster; this is akin to Rc.

The layout of this trait is also mandated to be that of a usize, for it is used for reference counting.