Trait tendril::ReadExt

pub trait ReadExt: Read {
    fn read_to_tendril<A>(
        &mut self, 
        buf: &mut Tendril<Bytes, A>
    ) -> Result<usize>
    where
        A: Atomicity;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Extension trait for io::Read.

Required Methods

fn read_to_tendril<A>(&mut self, buf: &mut Tendril<Bytes, A>) -> Result<usize> where
A: Atomicity,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?