Trait tendril::SliceExt

pub trait SliceExt<F>: Slice where
    F: SliceFormat<Slice = Self>, {
    fn to_tendril(&self) -> Tendril<F> { ... }
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Tendril-related methods for Rust slices.

Provided Methods

fn to_tendril(&self) -> Tendril<F>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Make a Tendril from this slice.