Trait tendril::fmt::CharFormat

pub unsafe trait CharFormat<'a>: Format {
    type Iter: Iterator<Item=(usize, char)>;
    unsafe fn char_indices(buf: &'a [u8]) -> Self::Iter;
    fn encode_char<F>(ch: char, cont: F) -> Result<(), ()> where F: FnOnce(&[u8]);
}

Indicates a format which contains characters from Unicode (all of it, or some proper subset).

Associated Types

type Iter: Iterator<Item=(usize, char)>

Iterator for characters and their byte indices.

Required Methods

unsafe fn char_indices(buf: &'a [u8]) -> Self::Iter

Iterate over the characters of the string and their byte indices.

You may assume the buffer is already validated for Format.

fn encode_char<F>(ch: char, cont: F) -> Result<(), ()> where F: FnOnce(&[u8])

Encode the character as bytes and pass them to a continuation.

Returns Err(()) iff the character cannot be represented.