Struct tendril::stream::LossyDecoder

pub struct LossyDecoder<Sink, A = NonAtomic> where
    Sink: TendrilSink<UTF8, A>,
    A: Atomicity,  {
    inner: LossyDecoderInner<Sink, A>,
}

A TendrilSink adaptor that takes bytes, decodes them as the given character encoding, lossily replace ill-formed byte sequences with U+FFFD replacement characters, and emits Unicode (StrTendril).

This allocates new tendrils for encodings other than UTF-8.

Fields
inner: LossyDecoderInner<Sink, A>
Implementations

impl<Sink, A> LossyDecoder<Sink, A> where
Sink: TendrilSink<UTF8, A>,
A: Atomicity,

pub fn new_encoding_rs(encoding: &'static Encoding, sink: Sink) -> Self

Create a new incremental decoder using the encoding_rs crate.

pub fn utf8(sink: Sink) -> LossyDecoder<Sink, A>

Create a new incremental decoder for the UTF-8 encoding.

This is useful for content that is known at run-time to be UTF-8 (whereas Utf8LossyDecoder requires knowning at compile-time.)

pub fn inner_sink(&self) -> &Sink

Give a reference to the inner sink.

pub fn inner_sink_mut(&mut self) -> &mut Sink

Give a mutable reference to the inner sink.