Struct tendril::Tendril

pub struct Tendril<F> where F: Format {
    // some fields omitted
}

Compact string type for zero-copy parsing.

Tendrils have the semantics of owned strings, but are sometimes views into shared buffers. When you mutate a Tendril, an owned copy is made if necessary. Further mutations occur in-place until the string becomes shared, e.g. with clone() or subtendril().

Buffer sharing is accomplished through thread-local (non-atomic) reference counting, which has very low overhead. The Rust type system will prevent you at compile time from sending a Tendril between threads. We plan to relax this restriction in the future; see README.md.

Whereas String allocates in the heap for any non-empty string, Tendril can store small strings (up to 8 bytes) in-line, without a heap allocation. Tendril is also smaller than String on 64-bit platforms — 16 bytes versus 24.

The type parameter F specifies the format of the tendril, for example UTF-8 text or uninterpreted bytes. The parameter will be instantiated with one of the marker types from tendril::fmt. See the StrTendril and `ByteTendril' type aliases for two examples.

The maximum length of a Tendril is 4 GB. The library will panic if you attempt to go over the limit.

Methods

impl<F> Tendril<F> where F: Format

fn new() -> Tendril<F>

Create a new, empty Tendril in any format.

fn with_capacity(capacity: u32) -> Tendril<F>

Create a new, empty Tendril with a specified capacity.

fn len32(&self) -> u32

Get the length of the Tendril.

This is named not to conflict with len() on the underlying slice, if any.

fn is_shared(&self) -> bool

Is the backing buffer shared?

fn is_shared_with(&self, other: &Tendril<F>) -> bool

Is the backing buffer shared with this other Tendril?

fn clear(&mut self)

Truncate to length 0 without discarding any owned storage.

fn try_from_byte_slice(x: &[u8]) -> Result<Tendril<F>, ()>

Build a Tendril by copying a byte slice, if it conforms to the format.

fn as_bytes(&self) -> &Tendril<Bytes>

View as uninterpreted bytes.

fn into_bytes(self) -> Tendril<Bytes>

Convert into uninterpreted bytes.

fn as_superset<Super>(&self) -> &Tendril<Super> where F: SubsetOf<Super>, Super: Format

View as a superset format, for free.

fn into_superset<Super>(self) -> Tendril<Super> where F: SubsetOf<Super>, Super: Format

Convert into a superset format, for free.

fn try_as_subset<Sub>(&self) -> Result<&Tendril<Sub>, ()> where Sub: SubsetOf<F>

View as a subset format, if the Tendril conforms to that subset.

fn try_into_subset<Sub>(self) -> Result<Tendril<Sub>, Tendril<F>> where Sub: SubsetOf<F>

Convert into a subset format, if the Tendril conforms to that subset.

fn try_as_other_format<Other>(&self) -> Result<&Tendril<Other>, ()> where Other: Format

View as another format, if the Tendril conforms to that format.

fn try_into_other_format<Other>(self) -> Result<Tendril<Other>, Tendril<F>> where Other: Format

Convert into another format, if the Tendril conforms to that format.

fn try_push_bytes(&mut self, buf: &[u8]) -> Result<(), ()>

Push some bytes onto the end of the Tendril, if they conform to the format.

fn push_tendril(&mut self, other: &Tendril<F>)

Push another Tendril onto the end of this one.

fn try_subtendril(&self, offset: u32, length: u32) -> Result<Tendril<F>, SubtendrilError>

Attempt to slice this Tendril as a new Tendril.

This will share the buffer when possible. Mutating a shared buffer will copy the contents.

The offset and length are in bytes. The function will return Err if these are out of bounds, or if the resulting slice does not conform to the format.

fn subtendril(&self, offset: u32, length: u32) -> Tendril<F>

Slice this Tendril as a new Tendril.

Panics on bounds or validity check failure.

fn try_pop_front(&mut self, n: u32) -> Result<(), SubtendrilError>

Try to drop n bytes from the front.

Returns Err if the bytes are not available, or the suffix fails validation.

fn pop_front(&mut self, n: u32)

Drop n bytes from the front.

Panics if the bytes are not available, or the suffix fails validation.

fn try_pop_back(&mut self, n: u32) -> Result<(), SubtendrilError>

Drop n bytes from the back.

Returns Err if the bytes are not available, or the prefix fails validation.

fn pop_back(&mut self, n: u32)

Drop n bytes from the back.

Panics if the bytes are not available, or the prefix fails validation.

unsafe fn as_other_format_without_validating<Other>(&self) -> &Tendril<Other> where Other: Format

View as another format, without validating.

unsafe fn into_other_format_without_validating<Other>(self) -> Tendril<Other> where Other: Format

Convert into another format, without validating.

unsafe fn from_byte_slice_without_validating(x: &[u8]) -> Tendril<F>

Build a Tendril by copying a byte slice, without validating.

unsafe fn push_bytes_without_validating(&mut self, buf: &[u8])

Push some bytes onto the end of the Tendril, without validating.

unsafe fn unsafe_subtendril(&self, offset: u32, length: u32) -> Tendril<F>

Slice this Tendril as a new Tendril.

Does not check validity or bounds!

unsafe fn unsafe_pop_front(&mut self, n: u32)

Drop n bytes from the front.

Does not check validity or bounds!

unsafe fn unsafe_pop_back(&mut self, n: u32)

Drop n bytes from the back.

Does not check validity or bounds!

impl<F> Tendril<F> where F: SliceFormat

fn from_slice(x: &Slice) -> Tendril<F>

Build a Tendril by copying a slice.

fn push_slice(&mut self, x: &Slice)

Push a slice onto the end of the Tendril.

impl Tendril<UTF8>

fn pop_front_char(&mut self) -> Option<char>

Remove and return the first character, if any.

fn pop_front_char_run<C, R>(&mut self, classify: C) -> Option<(Tendril<UTF8>, R)> where C: FnMut(char) -> R, R: PartialEq

Remove and return a run of characters at the front of the Tendril which are classified the same according to the function classify.

Returns None on an empty string.